Update open access publication query
This does two things. First, it considers the new
scholarsphere_open_access_url field in the query since publications that
now have a value in this field are open access. Second, it ignores empty
string values. Whenever a record is saved in the admin interface, any
empty form fields that are shown in the UI will end up replacing `nil`
values with empty strings in the database. As a result we have a lot of
publications in the production database where the open_access_url field
contains an empty string, and we definitely don’t want to count those
as open access publications.