Remove button/inline form for creating users for external pub. waivers
This form doesn’t work here because it seems to take its configuration
from that of the rails_admin `edit` action for users rather than from
the `create` action as you would expect. Editing the WebAccess ID field
for users is disabled/prohibited, but it’s also a required field for
creating a new user. This combination of business rules and what appears
to be a bug in rails_admin’s application of the action configuration
mean that it’s impossible to save a new user from this form without
validation errors.