Update Activity Insight publication authorship importing logic
This essentially changes two things. First, from a given user’s set of
metadata from Activity Insight, we’re only going to import their own
authorship record for each of their publications. We know that some of
the other authorship that are listed may be incorrect, so we’re going to
depend on any other authors listing the same publication in AI and the
duplicate publications eventually being merged in RMD to get the correct
full list of authors. Second, we’re no longer deleting authorship
records that are no longer represented in Activity Insight. We no longer
see the need to do this since it was originally just a half-measure to
address the problem of incorrect authorships in AI and here we’ve
updated the import process so that incorrect authorships should no
longer get imported. We’re also planning to delete all of the AI
authorship records that have been imported in the past and to reimport
that data in accordance with these new rules.