Abbreviate first names of other Penn State faculty authors
When we’re importing authorship metadata for a given user’s publications
from Activity Insight, we know that the first names of the authors that
are listed in the metadata are sometimes incorrect, but we know that the
first initial should always be correct. 

This is the case because of the way that some metadata is imported into
Activity Insight from other sources. Those other sources sometimes list
authors by first initial and last name. When an Activity Insight user is
importing their publication metadata from these other sources, AI
searches for other users in its database based on the names of the
authors provided by the external sources. Especially since those names
sometimes include just first initials, AI often matches those names to
other Penn State faculty members incorrectly, thus misattributing the
publication to those faculty members. As a result, we can be sure of the
following:

1. If the authorship representation that we’re importing is for the
   the person who entered the data into Activity Insight, then the full
   first and last name of the author should be correct.
2. If the authorship representation that we’re importing is not
   connected to any user in Activity Insight, then the full first and
   last name of the author should also be correct.
3. If the authorship representation that we’re importing is connected to
   some user in Activity Insight who is *not* the person who entered the
   data into Activity Insight, then the last name should be correct, and
   the first initial should be correct, but the full first name may be
   incorrect.

We want to import exactly the metadata that we are reasonably sure is
correct - no more and no less.