Create a new task for importing AI authorships
This code is intended to be run once in production after a subset of
the Authorship records that have already been imported from Activity
Insight has been deleted. Once the task has been run, this code can
be removed from the app.

The existing records need to be deleted because some of them are
incorrect (but we don’t know exactly which ones). Our modified import
process should then reimport only Authorship records from Activity
Insight that we know to be correct.