Check all publications for OAB import regardless of recent checks
At least for now, we want to ensure that we have as much of the
available open access information as possible right before we send out
open access reminder emails to faculty. We don’t want to send a reminder
to someone about one of their publications if the publication was just
recently added to Open Access Button. We may want to toggle this change
in the future depending on the desired schedule for Open Access Button
importing in general and prior to email sending.