Add confirmation to the “Upload to Scholarsphere” button
For faculty users, it’s not clear what the effects of clicking this
button will be. Thus it requires some explanation/confirmation so that
they don’t do something that they didn’t intend to do and so that they
aren’t surprised by the result or left wondering what to do next.