Cascade publication deletion to associated import records
Once in a while a publication is incorrectly listed in one of RMD’s
import sources for publication metadata. In those cases, RMD admins
work to remove the incorrect metadata from the source, but then they
also need to remove it from RMD. However, this was impossible previously
since publications with imports couldn’t be deleted due to foreign key
constraints, and there was also no way to delete import records
directly. Import records shouldn’t normally be deleted on their own, so
rather than provide the ability to delete them separately, I decided to
cascade the deletion of publications to associated imports so that they
can all be deleted at once.