Remove default limit for user/publication query
This default limit was causing bugs in the user profile API/logic/
presentation for any users that had more than 100 publications. Since
this default limit isn’t tested or documented anywhere, and since I
can’t see any clear need for it, I’m just going to remove it. If we
determine that we do want/need a default limit here in the future, then
it would be nice to have some way to override the limit feature entirely
so that the query returns unlimited results (even if the override isn’t
documented in the public REST API).