Automatically make publications resulting from duplicate merges visible
In some cases, publications may have been automatically hidden when they
were grouped with possible duplicates so that duplicates wouldn’t show
up in API responses or user profiles prior to the duplication being
resolved. Normally, once a duplicate group is resolved/merged, we’d want
the remaining publication to become visible if it was not already. Up to
this point, we had been manually changing these visibility settings in
bulk after many duplicate publication groups had been merged, but it
makes sense to automate this change as part of the duplicate merging
process.