Tweak the data model for Scholarsphere deposits
We may need a place to store some user-provided metadata that is
specific to ScholarSphere while we’re waiting for the deposit to be
made. However, even if we don’t, I think that it’s going to be nicer to
record and track the status of ScholarSphere deposits using a separate
database table rather than adding a bunch of unrelated fields to the
authorships table.