Add metadata fields to ScholarSphere work deposit records
These are fields that are required by ScholarSphere in order to create
and publish a work via the API. Typically the title, description, and
date fields will be filled in with metadata that already exists in RMD,
but we’ll give the user the opportunity to edit that data if they want
before it’s sent to ScholarSphere. If any of that metadata is missing
from RMD, then the user will need to provide it in order to publish in
ScholarSphere.