Don’t retry failed background jobs
In particular, the job for depositing works in ScholarSphere is not
idempotent, and it changes state both in RMD and in ScholarSphere. So,
in the event of an unexpected error, it’s possible for a job to
partially run before failing. The best way to handle this right now
might be to give up on a failed job after one try and keep the failed
job in the database for inspection. That should allow us to determine
what state everything is in after a job failure, how it might need to be
corrected, and how to prevent the same kind of failure from happening
again.