Handle ScholarSphere work deposit failures
There are generally two kinds of ways in which a user’s attempt to
deposit a work in ScholarSphere from RMD might fail. First, every part of
the process, including sending the request to ScholarSphere may succeed,
but ScholarSphere may return a non-200 HTTP response. This would be
unexpected because any non-200 response from ScholarSphere would be due
to something that is beyond the RMD user’s control. A non-200 response
would therefore indicate a problem with RMD’s integration with
ScholarSphere such as RMD not properly validating the user’s input
before sending it to ScholarSphere. The other mode of failure would be
if the RMD code that performs the ScholarSphere submission raises an
error. This might happen, for example, if the network connection
between RMD and ScholarSphere fails.

In either case, any failure would be both unanticipated and difficult to
properly recover from. So we ultimately handle all failures in
essentially the same way. In the case of a non-success response from
ScholarSphere, we update the status of the deposit record to ‘Failed’
and record the error message from the ScholarSphere response (note that
this counts as a “successful” background job, so in this case, the job
will be completed and removed from the queue). In the case of the app
raising an exception, we rescue, again record the status and error
message in the deposit record, and then re-raise. In this case, the
exception causes the background job to fail, so the failed job record
remains in the database in addition to the update to the deposit record.

Once the status of the deposit is ‘Failed’, a user can see this in the
publication list in the profile management interface. At this point,
they’re also able to attempt the deposit again from scratch. If a user
remains unable to successfully make a deposit, we’ll potentially see
clues as to the reason in the resulting ScholarsphereWorkDeposit
records, Active Job records, and/or the scholarsphere_deposit.log file.
File uploads for failed deposits won’t be cleaned up, so once the source
of a problem is addressed, it may be possible for an RMD developer to
retry a failed deposit on a user’s behalf, but any such retry should be
done with care and the awareness that the deposit process itself is not
idempotent.