Use separate queues for ScholarSphere upload jobs
The name of each queue includes the name of the host on which the jobs
were enqueued. The idea is that if there is more than one host per
environment, then on each host, delayed job will be configured to only
run jobs from the queue for that specific host. This may be necessary at
first if multiple hosts don’t share the same filesystem where the
uploaded files are stored. If each host stores uploaded files in its own
separate filesystem, then it can only run ScholarSphere upload jobs that
it enqueued. Assuming that we ever get to the point where we’re using
multiple hosts, we may obviate the need for this by mounting a shared
filesystem to store uploaded files. In that case, we’d probably still
want a separate queue for the ScholarSphere upload jobs, but we’d only
need one, and it wouldn’t need to include a hostname.