Wrap delayed_job in a helper script
For the delayed_job script to affect the processes that it started,
subsequent commands need to be given the same arguments that were given
along with the `start` command. It’s easy to forget that this is the
case and to issue commands that don’t do work as intended. The purpose
of this script is to encapsulate the options/arguments and ensure that
they’re passed consistently so that regardless of whether the script is
being run by an init script on a server, by an administrator, etc., the
same options will always be applied for every command.