Load ScholarSphere client config into environment variables
This is a little bit of a stab in the dark. The ScholarSphere client
should be doing this itself, but when loaded in the context of a
delayed_job process (and only in a deployed environment), it’s not
happening for some reason. I’m trying to see if this helps.