Don’t roll back other updates if file cleanup fails
The ScholarSphere deposit is still a “success” even if we fail to clean
up the temporary storage of the file uploads. We wouldn’t want any error
raised by the file clean-up to prevent the deposit’s associated
publication from getting updated with the open access URL from
ScholarSphere since presumably the files and metadata were actually
deposited in ScholarSphere, and we wouldn’t want the user to think that
the deposit failed. If this happened, the user may try to do the same
(actually successful) deposit again, thereby creating a duplicate
deposit in ScholarSphere. However, we do want to know if the file
clean-up failed so that we can do something about that problem. We
should still find out about this kind of failure because an error raised
here will later be caught and the deposit’s status will be changed from
‘Success’ to ‘Failed’ with the associated error message. While the
status will be slightly misleading, the error message, and the fact that
the publication now has an open access URL from ScholarSphere should
clarify what went wrong.