Add DOI field to ScholarSphere work deposits
I made the decision to make the form field disabled if there is already
a DOI present from the RMD publication metadata. This allows a user to
enter a DOI if there isn’t one in RMD, but they can’t change a DOI that
we already have. This seems good for the integrity of the data going
into ScholarSphere, but I’m not certain that it’s the right behavior. I
don’t know if there will be any cases where we have a DOI that’s somehow
incorrect, but if such cases exist, then it’s probably better that the
user contacts RMD support and actually gets the problem sorted out in
RMD so that we have accurate DOIs.