List the “creators” that will be submitted to ScholarSphere
We’re not giving users the ability to change this information at this
point, but I think that they should be able to see everything that is
being submitted to ScholarSphere on their behalf. If something here is
somehow incorrect, we’d want to know about it, and if the user just 
wants to make a superficial change to how the names are presented, then
they could edit the work after it has been created in ScholarSphere.