Transfer ScholarSphere deposit records whenever authorships are merged
The need for this is kind of hypothetical at this point, but it’s
technically possible that two or more authorship records that need to be
merged together could each have one or more associated ScholarSphere
work deposit records. If this did happen, it would probably be the
result of some sort of mistake, so it would be nice to keep all of the
records around to help us determine exactly what happened rather than
just deleting the deposit records that are associated with the
deleted authorships. Keeping all of the records would provide a full
history of every ScholarSphere interaction involving the publication,
and it won’t affect the status of the publication at this point.