Create a task that imports URLs for previously unchecked publications
We’ve noticed that there are publication records in the database that
meet all of the criteria for their open access metadata to be imported
from Open Access Button, and yet it’s not happening when we run the full
Open Access Button import. At this point, we aren’t sure why this is the
case, so we want to run a much smaller import that attempts to get the
metadata for only those publications that meet the criteria but have not
been looked up in Open Access Button previously. We’re hoping that this
will both import the open access metadata that we’re missing and help
to shed some light on why it was missing in the first place.