Fix some problematic DOI values
Some DOIs with these problematic characters are somehow ending up in the
database. We’re not sure of the source(s), but these DOI values are
causing the Open Access Button import to produce errors that prevent the
import from completing. We’ll probably need to apply this kind of
filtering to any point where publication DOIs are saved to the database.