Output some useful information for commonly encountered OAB import error
A TypeError occurs here when we get a valid JSON response back from the
Open Access Button API after looking up a publication by DOI, but the
contents of the response are some kind of error message rather than the
metadata that we’re expecting. This situation can occur for many
different reasons (some of which may be totally unexpected), so any
automatic error handling is going to be difficult. Being able to see
some context for the errors will allow us to handle them manually, at
least for now.