Validate publication DOI format before saving
We use the same logic to sanitize publication DOIs as they are imported.
So all DOIs that are coming from data imports should pass this
validation, but it’s also possible for admin users to add/update DOIs
via the admin UI, and it’s very important for us to have only valid DOIs
that are in a consistent format since they’re used to compare
publications for uniqueness and to look up publication metadata
elsewhere.