Temporarily force reimporting of all phone numbers from Activity Insight
We have many incorrect values in the database due to the way that we
were previously storing the phone number segments (as integers rather
than strings). Since it’s unlikely that any phone number data has been
manually updated by admin users in RMD, we just want to reimport and
overwrite all of the AI phone number data in order to correct all of the
erroneous values. This change should be reverted after we run one AI
import.