Use Selenium webdrivers in place of poltergeist
Headless chrome drivers are a replacement for PhantomJS, which is now
outdated. This removes the need for the phantomjs local dependency and
uses Chrome as its browser.

In order to test downloads, a configured download directory is added
to the testing browser, which we can monitor for downloads to verify if
they have completed successfully.