Add Penn State identity information to users
Integrates the Penn State identity management API into the User record.
The response from the API is stored as json in a column of the user
table. When retrieved, a PsuIdentity::SearchService::Person is returned,
containing additional metadata about the user such as their preferred
names, email, and affiliations.

This data can be used to determine if the user is still active at Penn
State.