Add a timestamp to psu_identity; update via UserProfile
Even if a user's psu identity is nil, we should still indicate that
we've updated it with the API. This would cover instances of existing
users that leave the University, and then their API data changes to nil,
or other users that we haven't been able to verify yet.

The UserProfile can coordinate identity data via #active?. When a
profile is built, it will check with the API if the identity data is nil
and it hasn't been checked yet.