Remove OAB-sourced metadata that has been removed from OAB
This importer was naively assuming that once Open Access Button reports
that a publication is open access, then it will always be open access.
In fact, publications can change from being open access to being closed,
and the Open Access Button data will reflect this change by providing
metadata that does not include an open access URL where one had been
included before. We definitely want to record this change in status
because if the PSU open access policy applies to the publication, then
we’ll want to try to make it open access again.