Create a DeputyAssignment to assign delegate users
A user (primary) may assign a delegate (deputy) in order to allow that
user to act on their behalf. The deputy may then upload articles,
submit waivers, or otherwise interact with the application as if they
were logged in as the primary user.

This creates a join table, deputy_assignments, that links a user with
their deputies, and notes whether or not they are active and if they
have confirmed their willingness to be this user's delegate.

Validations restrict delegates to non-admin users and prevent a user
from having the same delegate listed more than once.

Co-authored-by: Ryan Schenk <ryan@westarete.com>