Allow deputy users to masquerade as their primaries
When a primary user assigns another user as their deputy, they should be
able to masquerade as that user and do things on their behalf. This
feature already existed for admins, so we're expanding it to include
users that are assigned deputy roles from another primary user.

To preserve the existing feature, the original admin functions are
extracted to a shared module. The admin controller then only needs to
include the core functions of that module. The user controller uses
those same features but adds additional protections that ensure only the
deputy may become the primary.

Once the deputy is vetted, they become the primary user via
CurrentUserBuilder. However, certain circumstances may arise where the
primary user removes their deputy, revoking their privileges, but that
deputy may still have an active session. To guard against that, we
always check the deputies privileges for _each request_. The downside of
this is that we can't memoize current_user. CurrentUserBuilder is
always called, even if the user it not intending to masquerade as
another user.