Show more publications in the search list
Given the use case, we probably don’t need to go crazy with pagination
or further filtering of this list. We just want to make sure that a
hundred thousand results aren’t returned if a user does a title search
for “the”, but we also want to make sure that a user can see enough
results once they are able to search by author name in addition to
title.