Use test doubles in place of PSU identity service API wrapper
Due to the way that we’re fetching user metadata from Penn State’s
identity service whenever the user’s profile metadata is accessed for
the first time or whenever the user signs in, we had an enormous number
of tests that were incidentally causing real HTTP requests to be sent to
the identity service and were thus depending on its actual responses.
This was very undesirable for several reasons, not the least of which
being that nearly all of the integration tests would fail if the
identity service was down. Considering that none of the tests that were
doing this were actually intended to exercise the app’s interaction with
the real identity service, and considering that the identity service API
is hidden behind an interface provided by a gem that is presumably
thoroughly tested, my approach to fixing this was to mock the gem. We’re
hiding the gem behind our own interface, and the tests for that class
are exercising the gem but using canned HTTP responses in place of live
responses from the identity service API. Since it would be nice to
verify that this application is interacting correctly with the real
identity service and since we no longer have any tests that do that, I
think we should write one integration test for that specific purpose.
Then if there are breaking changes to the identity service API, we’ll
know, and if the API is down, we’ll only have one failing test instead
of hundreds.