Try a different strategy for downloading files from Activity Insight
We didn’t like the dependency on `wget` to begin with, and we discovered
that the behavior is inconsistent between our development environments
and the production environment. It was also somewhat awkward to try to
stream data from `wget` to CarrierWave for storage. This (hopefully)
eliminates all of those problems by using the Ruby standard library for
HTTP to download the file and store it in the same way that CarrierWave
would.