Construct query for publications that need manual permissions review
I believe that this does what we want, but we’re probably going to need
some real-world testing to be sure. The spec ensures that the right
records are being included in the result, but it could use many more
test records to ensure that we’re excluding all of the right things.