Update the scope for the final metadata review list in DM-OA workflow
This scope is now based on open access permissions metadata attributes
of the associated file records rather than the attributes of the
publication record, and it needs to account for publications where
either the published or accepted version can be deposited.

There is still test coverage at the model level for this scope, but it
can be improved.