WIP on specifying all correct behavior for merging publication DOI info
There are many possible distinct scenarios that may be encountered when
reconciling the DOI information in a group of publications that are
being merged together. We only accounted for a few of those scenarios
previously, and we discovered that there were real-world cases where the
behavior in those particular scenarios was both unspecified and
incorrect.

This is an attempt to specify all of the correct behavior before
actually correcting the behavior. There is still a little bit of work to
do to further specify a few scenarios where there is a “tie” between
differing DOI information with the same precedence, but this should be
enough to get started with correcting the behavior in most
circumstances.

Note that we’re fairly, but not entirely sure that we have specified the
behavior that we actually want in every single case, so it’s possible
that the specification could change, but at this point it should at
least be complete.