Design adjustments to incorporate changes to shared assets
There are recent changes to the assets in psulib_base, which we point to.
For the most part, everything looks good, but one of our links (which is actually a button)
became difficult to read because of low color contrast.
This commit makes the button text in our nav bar match that of our links for consistency and legibility.